Fix tests after rebase

+++ unit-tests/src/test/scala/scala/scalanative/unsafe/PtrBoxingTest.scala
@@ -6,7 +6,7 @@ import org.junit.Assert._
 
 import scalanative.junit.utils.AssertThrows._
 
-import scalanative.runtime.toRawPtr
+import scalanative.unsigned._
 import scalanative.libc.stdlib.malloc
 import java.lang.Long.toHexString
 
@@ -15,8 +15,8 @@ class PtrBoxingTest {
   var any: Any = null
 
   @noinline lazy val nullPtr: Ptr[Byte] = null
-  @noinline lazy val ptr: Ptr[Byte]     = malloc(64)
-  @noinline lazy val ptr2: Ptr[Byte]    = malloc(64)
+  @noinline lazy val ptr: Ptr[Byte]     = malloc(64.toULong)
+  @noinline lazy val ptr2: Ptr[Byte]    = malloc(64.toULong)
 
   @noinline def f[T](x: T): T      = x
   @noinline def cond(): Boolean    = true
